Remove old date validator. Rename newDateValidator to dateValidator. Add validation that the date doesn't contain non-numeric characters. Update date validator tests.'
