+++ server/constants/strings.ts
@@ -7,9 +7,7 @@ const strings: Record<string, Record<string, string>> = {
     missingDay: 'Date of birth must include a day',
     missingMonth: 'Date of birth must include a month',
     missingYear: 'Date of birth must include a year',
-    unrealDate: 'Please enter a real date',
-    incorrectDateFormat:
-      'Date is in the incorrect format. Enter the date in the format DD/MM/YYYY (Day/Month/Year). For example, 24/10/2020',
+    unrealDate: 'Please enter a real date in the format DD/MM/YYYY. For example, 24/10/2020',
     dateBelowLimit: 'Enter a date after ',
     futureDate: 'Future dates are not allowed',
   },
