+++ server/services/datastoreSearchService.ts
@@ -6,7 +6,7 @@ import DatastoreClient from '../data/datastoreClient'
 import { HmppsAuthClient, RestClientBuilder } from '../data'
 import { ValidationResult } from '../models/Validation'
 import { SearchFormInput, SearchResultsRequest } from '../types/Search'
-import { DateValidationResponse, dateValidator } from '../utils/validators/newDateValidator'
+import { DateValidationResponse, dateValidator } from '../utils/validators/dateValidator'
 import NameValidator from '../utils/validators/nameValidator'
 import { ParsedSearchFormData } from '../models/form-data/searchOrder'
 import { QueryExecutionResponse } from '../interfaces/QueryExecutionResponse'
